LayoutTests/http/wpt/webauthn/public-key-credential-abort-reason.https.html

<!DOCTYPE html><!-- webkit-test-runner [ WebAuthenticationModernEnabled=true allowTestOnlyIPC=true ] -->
<title>Web Authentication API: AbortSignal reason is propagated on credentials.create() and credentials.get()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/util.js"></script>
<script>
    // Use a HID mock that responds with an error so the callback fires and
    // the abort-reason check in AuthenticatorCoordinator is exercised.
    if (window.internals)
        internals.setMockWebAuthenticationConfiguration({ hid: { stage: "request", subStage: "msg", error: "malicious-payload" } });

    // credentials.create() — signal aborted before call

    promise_test(async t => {
        const controller = new AbortController();
        controller.abort("CustomReason");

        const options = {
            publicKey: {
                rp: { name: "example.com" },
                user: { name: "John Appleseed", id: asciiToUint8Array("123456"), displayName: "John" },
                challenge: asciiToUint8Array("123456"),
                pubKeyCredParams: [{ type: "public-key", alg: -7 }],
            },
            signal: controller.signal,
        };

        await navigator.credentials.create(options).then(
            t.unreached_func("Should have rejected"),
            reason => assert_equals(reason, "CustomReason", "Should reject with signal.reason, not a generic AbortError")
        );
    }, "credentials.create() rejects with signal.reason when signal is pre-aborted with a string reason");

    promise_test(async t => {
        const controller = new AbortController();
        const error = new DOMException("Custom abort", "AbortError");
        controller.abort(error);

        const options = {
            publicKey: {
                rp: { name: "example.com" },
                user: { name: "John Appleseed", id: asciiToUint8Array("123456"), displayName: "John" },
                challenge: asciiToUint8Array("123456"),
                pubKeyCredParams: [{ type: "public-key", alg: -7 }],
            },
            signal: controller.signal,
        };

        await navigator.credentials.create(options).then(
            t.unreached_func("Should have rejected"),
            reason => assert_equals(reason, error, "Should reject with the exact DOMException passed to abort()")
        );
    }, "credentials.create() rejects with signal.reason when signal is pre-aborted with a DOMException");

    // credentials.get() — signal aborted before call

    promise_test(async t => {
        const controller = new AbortController();
        controller.abort("CustomReason");

        const options = {
            publicKey: {
                challenge: asciiToUint8Array("123456"),
            },
            signal: controller.signal,
        };

        await navigator.credentials.get(options).then(
            t.unreached_func("Should have rejected"),
            reason => assert_equals(reason, "CustomReason", "Should reject with signal.reason, not a generic AbortError")
        );
    }, "credentials.get() rejects with signal.reason when signal is pre-aborted with a string reason");

    promise_test(async t => {
        const controller = new AbortController();
        const error = new DOMException("Custom abort", "AbortError");
        controller.abort(error);

        const options = {
            publicKey: {
                challenge: asciiToUint8Array("123456"),
            },
            signal: controller.signal,
        };

        await navigator.credentials.get(options).then(
            t.unreached_func("Should have rejected"),
            reason => assert_equals(reason, error, "Should reject with the exact DOMException passed to abort()")
        );
    }, "credentials.get() rejects with signal.reason when signal is pre-aborted with a DOMException");
</script>
