LayoutTests/http/wpt/webauthn/public-key-credential-abort-reason.https-expected.txt


PASS credentials.create() rejects with signal.reason when signal is pre-aborted with a string reason
PASS credentials.create() rejects with signal.reason when signal is pre-aborted with a DOMException
PASS credentials.get() rejects with signal.reason when signal is pre-aborted with a string reason
PASS credentials.get() rejects with signal.reason when signal is pre-aborted with a DOMException
